pkg/cerror/middleware_test.go
//go:build unit

package cerror

import (
	"net/http"
	"testing"

	"github.com/gofiber/fiber/v2"
	"github.com/stretchr/testify/assert"
	"github.com/stretchr/testify/require"
	"go.uber.org/zap"
)

func TestMiddleware(t *testing.T) {
	t.Run("when custom error type error pass to middleware should resolve and return it", func(t *testing.T) {
		app := fiber.New(fiber.Config{
			ErrorHandler: Middleware,
		})
		app.Get("/", func(ctx *fiber.Ctx) error {
			return NewError(
				fiber.StatusInternalServerError,
				"something went wrong",
				zap.String("key", "value"),
			)
		})

		req, err := http.NewRequest(fiber.MethodGet, "/", nil)
		require.NoError(t, err)

		resp, err := app.Test(req)
		require.NoError(t, err)

		assert.Equal(t, fiber.StatusInternalServerError, resp.StatusCode)
	})

	t.Run("when error pass to middleware should resolve and return it", func(t *testing.T) {
		app := fiber.New(fiber.Config{
			ErrorHandler: Middleware,
		})
		app.Get("/", func(ctx *fiber.Ctx) error {
			return fiber.NewError(fiber.StatusInternalServerError, "something went wrong")
		})

		req, err := http.NewRequest(fiber.MethodGet, "/", nil)
		require.NoError(t, err)

		resp, err := app.Test(req)
		require.NoError(t, err)

		assert.Equal(t, fiber.StatusInternalServerError, resp.StatusCode)
	})
}
